Add CWD_PATH and set_cwd(path)
Setting the current working directory in which the Pandoc command is
executed allows for processing pdf and epub from a location other than
the cwd of the python code calling Pandoc.

When generating PDF output from markdown that contains:

  ![Figure 1](images/fig1.png)

Pandoc expects to be executed from the location containing the
'images' directory. When using pyandoc to generate a pdf from
source in a directory other than the cwd of your python code, use
'pandoc.set_cwd(src_path)' before running the Document.to_file() method.

In addition, for proper epub generation of the same file, Pandoc appears
to expect that the 'images' directory exists in the same location as
the markdown file being processed. Therefore the temporary markdown
file generated by Document.to_file() needs to be created in the same
location as the 'images' directory so dir=CWD_PATH is not passed to to
NamedTemporaryFile. As with pdf generation, pandoc.set_cwd(path) should
be called before calling Document.to_file().

By default CWD_PATH is set to '.'.